Refactor utils
r


fix
