+++ cve_bin_tool/async_utils.py
@@ -30,40 +30,24 @@ def get_event_loop():
     except RuntimeError:
         loop = asyncio.new_event_loop()
         asyncio.set_event_loop(loop)
-    if sys.platform.startswith("win"):
-        if isinstance(loop, asyncio.SelectorEventLoop):
-            loop = asyncio.ProactorEventLoop()
-            asyncio.set_event_loop(loop)
+    if sys.platform.startswith("win") and isinstance(loop, asyncio.SelectorEventLoop):
+        loop = asyncio.ProactorEventLoop()
+        asyncio.set_event_loop(loop)
     return loop
 
 
 def run_coroutine(coro):
     loop = get_event_loop()
     aws = asyncio.ensure_future(coro)
-    result = loop.run_until_complete(aws)
-    return result
+    return loop.run_until_complete(aws)
 
 
 async def aio_run_command(args):
     process = await asyncio.create_subprocess_exec(
-        *args,
-        stdout=asyncio.subprocess.PIPE,
-        stderr=asyncio.subprocess.PIPE,
+        *args, stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE
     )
     stdout, stderr = await process.communicate()
     return stdout, stderr  # binary encoded
-
-
-class ChangeDirContext:
-    def __init__(self, destination_dir):
-        self.current_dir = os.getcwd()
-        self.destination_dir = destination_dir
-
-    async def __aenter__(self):
-        os.chdir(self.destination_dir)
-
-    async def __aexit__(self, exc_type, exc_val, exc_tb):
-        os.chdir(self.current_dir)
 
 
 class FileIO:
