+++ cve_bin_tool/extractor.py
@@ -17,12 +17,12 @@ from .async_utils import (
     aio_rmdir,
     aio_makedirs,
     aio_mkdtemp,
-    ChangeDirContext,
     FileIO,
     run_coroutine,
 )
 from .error_handler import ExtractionFailed, UnknownArchiveType, ErrorHandler, ErrorMode
 from .log import LOGGER
+from .util import ChangeDirContext
 
 
 class BaseExtractor:
@@ -50,10 +50,10 @@ class BaseExtractor:
 
     def can_extract(self, filename):
         """ Check if the filename is something we know how to extract """
-        for extension in itertools.chain(*self.file_extractors.values()):
-            if filename.endswith(extension):
-                return True
-        return False
+        return any(
+            filename.endswith(extension)
+            for extension in itertools.chain(*self.file_extractors.values())
+        )
 
     @staticmethod
     async def extract_file_tar(filename, extraction_path):
@@ -173,7 +173,7 @@ class TempDirExtractorContext(BaseExtractor):
                     if os.path.exists(extracted_path):
                         await aio_rmdir(extracted_path)
                     await aio_makedirs(extracted_path, 0o700)
-                    async with ChangeDirContext(extracted_path):
+                    with ChangeDirContext(extracted_path):
                         if await extractor(filename, extracted_path) != 0:
                             if self.raise_failure:
                                 with ErrorHandler(
