+++ cve_bin_tool/util.py
@@ -112,6 +112,18 @@ def inpath(binary):
     )
 
 
+class ChangeDirContext:
+    def __init__(self, destination_dir):
+        self.current_dir = os.getcwd()
+        self.destination_dir = destination_dir
+
+    def __enter__(self):
+        os.chdir(self.destination_dir)
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        os.chdir(self.current_dir)
+
+
 class DirWalk:
     """
     for filename in DirWalk('*.c').walk(roots):
@@ -178,7 +190,4 @@ class DirWalk:
         """ Match filename patterns """
         if not patterns:
             return False
-        for pattern in patterns.split(";"):
-            if fnmatch.fnmatch(text, pattern):
-                return True
-        return False
+        return any(fnmatch.fnmatch(text, pattern) for pattern in patterns.split(";"))
